+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.3
 
 * Always put member metadata annotations on their own line (#483).
+* Indent functions in named argument lists with non-functions (#478).
 * Force the parameter list to split if a split occurs inside a function-typed
   parameter.
 * Don't force a split for before a single named argument if the argument itself